Repackage for re-release of 6.6.2
I mailed the bug-global mailing back in May about whether version 6.6.2
was re-released or not. The answer was yes, so I am repackaging 6.6.2
with the fixed tarball. Of course, the .sig file is different, because
the tarball changed.

See the following links for reference:
https://lists.gnu.org/archive/html/bug-global/2018-05/msg00000.html
https://lists.gnu.org/archive/html/bug-global/2018-05/msg00001.html

Also, switch to the mirrors.kernel.org GNU mirror.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>